Use linter config from OTEL Collector + Fix raised linter problems (#21)
* New linter configuration and fixes for all the linter errors

* fix test

* fix whitespace complaints

* Fix compilation of integration tests

* Fixed more test errors

* Revert 1.17 test changes

+++ pkg/internal/ebpf/generictracer/generictracer_test.go
@@ -25,5 +25,5 @@ func TestBitPositionCalculation(t *testing.T) {
 }
 
 func makeKey(first, second uint32) uint64 {
-	return uint64((uint64(first) << 32) | uint64(second))
+	return (uint64(first) << 32) | uint64(second)
 }
